move tests cases to unit
