+++ test/integration/test_serialization.py
@@ -12,13 +12,11 @@
 
 """Test serializing and deserializing data sent to the server."""
 
-import json
 from typing import Set, Any, Dict, Optional
 from unittest import SkipTest, skipIf
 
 import dateutil.parser
-from qiskit import transpile, schedule, assemble
-from qiskit import QuantumCircuit, QuantumRegister, ClassicalRegister
+from qiskit import transpile, schedule, QuantumCircuit
 from qiskit.circuit import Parameter
 from qiskit.test import slow_test
 from qiskit.test.reference_circuits import ReferenceCircuits
@@ -201,47 +199,6 @@ class TestSerialization(IBMTestCase):
         val = IBMJsonEncoder().default(param.bind({param: 0.2 + 0.1j}))
         self.assertEqual(val[0], 0.2)
         self.assertEqual(val[1], 0.1)
-
-    def test_exception_message(self):
-        """Test executing job with Parameter in methadata."""
-        quantum_register = QuantumRegister(1)
-        classical_register = ClassicalRegister(1)
-        my_circ_str = "test_metadata"
-        my_circ = QuantumCircuit(
-            quantum_register,
-            classical_register,
-            name=my_circ_str,
-            metadata={Parameter("φ"): 0.2},
-        )
-        qobj = assemble(my_circ)
-        qobj_dict = qobj.to_dict()
-        json.dumps(qobj_dict, cls=IBMJsonEncoder)
-        # There is no self.assert method because if we cannot pass Parameter as metadata
-        # the last line throw:
-        # "TypeError: keys must be str, int, float, bool or None, not Parameter"
-
-    def test_encode_no_replace(self):
-        """Test encode where there is no invalid key to replace."""
-        test_dir = {"t1": 1, None: None, "list": [1, 2, {"ld": 1, 2: 3}]}
-
-        self.assertEqual(
-            '{"t1": 1, "null": null, "list": [1, 2, {"ld": 1, "2": 3}]}',
-            IBMJsonEncoder().encode(test_dir),
-        )
-
-    def test_encode_replace(self):
-        """Test encode where there is no invalid key to replace."""
-        test_dir = {
-            "t1": 1,
-            None: None,
-            Parameter("a"): 0.2,
-            "list": [1, 2, {"ld": 1, 2: 3, Parameter("alfa"): 0.1}],
-        }
-
-        self.assertEqual(
-            '{"t1": 1, "null": null, "a": 0.2, "list": [1, 2, {"ld": 1, "2": 3, "alfa": 0.1}]}',
-            IBMJsonEncoder().encode(test_dir),
-        )
 
 
 def _find_potential_encoded(data: Any, c_key: str, tally: set) -> None:
